+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 ## v0.2.0 (Unreleased)
 
+### Features
+
+- **`Snex.Serde`: Rework serialization between Elixir and Python**
+
+  - iodata can be wrapped with `Snex.Serde.binary(b)` to efficiently pass it to Python out-of-band, without further encoding.
+    On the Python side, it's received as `bytes`.
+  - Python `bytes` are passed to Elixir through the same mechanism.
+  - Arbitrary Erlang terms can be wrapped with `Snex.Serde.term(t)`.
+    They will be encoded with `term_to_binary` and will be decoded in Python as opaque, tagged binaries.
+    Passing the value back to Elixir will decode it with `binary_to_term`.
+    Like raw binaries wrapped with `Snex.Serde.binary/1`, terms are efficiently passed out-of-band.
+  - Users can `defimpl Snex.Serde.Encoder` to define custom encoding of structs.
+    If no implementation is defined, encoding falls back to `JSON` encoding (`JSON.Encoder` and its defaults).
+    Implementations must encode structs into JSON.
+
 ### Changes
 
 - **Split Python script into multiple files**